fix: move to internal routes
