+++ openapi/mgmt/openapi.bundle.json
@@ -21,153 +21,6 @@
     }
   ],
   "paths": {
-    "/applications": {
-      "get": {
-        "operationId": "getApplications",
-        "tags": [
-          "Applications"
-        ],
-        "security": [
-          {
-            "ApiKeyAuth": []
-          }
-        ],
-        "summary": "List applications",
-        "description": "Get a list of applications",
-        "responses": {
-          "200": {
-            "description": "Applications",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "type": "array",
-                  "items": {
-                    "$ref": "#/paths/~1applications/post/responses/201/content/application~1json/schema"
-                  }
-                },
-                "examples": {
-                  "Example": {
-                    "value": [
-                      {
-                        "application_id": "your-applications-unique-application-id",
-                        "name": "My Staging App"
-                      },
-                      {
-                        "application_id": "your-other-applications-unique-application-id2",
-                        "name": "My Production App"
-                      }
-                    ]
-                  }
-                }
-              }
-            }
-          }
-        }
-      },
-      "post": {
-        "operationId": "createApplication",
-        "summary": "Create application",
-        "tags": [
-          "Applications"
-        ],
-        "security": [
-          {
-            "ApiKeyAuth": []
-          }
-        ],
-        "parameters": [],
-        "requestBody": {
-          "required": true,
-          "content": {
-            "application/json": {
-              "schema": {
-                "type": "object",
-                "properties": {
-                  "name": {
-                    "type": "string",
-                    "example": "My Production App"
-                  }
-                },
-                "required": [
-                  "name"
-                ]
-              }
-            }
-          }
-        },
-        "responses": {
-          "201": {
-            "description": "Application created",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "type": "object",
-                  "properties": {
-                    "id": {
-                      "type": "string",
-                      "example": "d8ceb3ff-8b7f-4fa7-b8de-849292f6ca69"
-                    },
-                    "name": {
-                      "type": "string",
-                      "example": "My Production App"
-                    }
-                  },
-                  "required": [
-                    "id",
-                    "name"
-                  ]
-                }
-              }
-            }
-          }
-        }
-      }
-    },
-    "/applications/{application_id}": {
-      "get": {
-        "operationId": "getApplication",
-        "summary": "Get application",
-        "tags": [
-          "Applications"
-        ],
-        "security": [
-          {
-            "ApiKeyAuth": []
-          }
-        ],
-        "responses": {
-          "200": {
-            "description": "Application",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "$ref": "#/paths/~1applications/post/responses/201/content/application~1json/schema"
-                },
-                "examples": {
-                  "Example": {
-                    "value": {
-                      "application_id": "d8ceb3ff-8b7f-4fa7-b8de-849292f6ca69",
-                      "name": "My App"
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      },
-      "parameters": [
-        {
-          "name": "application_id",
-          "in": "path",
-          "required": true,
-          "schema": {
-            "type": "string",
-            "example": "your-applications-unique-application-id"
-          }
-        }
-      ]
-    },
     "/customers": {
       "get": {
         "operationId": "getCustomers",
